[Perl] Attempting to veer toward more perl-natural implementation.
To be able to get *something* compact like conde, I'm using overloading
of the bitwise & and | operators.

To get our version of (fresh (x y) ...), we're currently passing the
number of fresh variables needed to call_fresh.  Kind of ugly.

However, as the tests show, it's becoming more natural.  It's fairly
usable (if verbose) right now.